28.1. Featured Sidebar
Here's our mission: to render a featured product on the cart sidebar with fully-functional "add to cart" controls. OoooOoOo.
Start by going to the asset/services/ directory and opening products-service.js. This file was included in the starting code for the tutorial and it already has a function called fetchFeaturedProducts. This returns a Promise that resolves to an Axios response that contains an array of products that our API says should be "featured". Phew!

 57 lines  assets/services/products-service.js
export function fetchFeaturedProducts() {
    return axios.get('/api/products', {
        params: { featured: 1 },
    });
}
Now go into shopping-cart.vue: the component that renders the entire cart page. Here's what I'm thinking: when this component is created, we make an AJAX request to fetch the featured products. We'll then find the first featured product in that collection, set it onto a new piece of data on this component and render it up here in the sidebar.
Let's do it!

Fetching the Featured Products
Start by adding the new piece of data: featuredProduct set to null.
To make the AJAX request, let's create a method and then call that method from created. How about: loadFeaturedProducts. Inside, say const featuredProducts = and execute that fetchFeaturedProducts() function we just saw. Make sure to auto-complete this so PhpStorm adds the import on top.
fetchFeaturedProducts returns a Promise whose data resolves to a response. So what we really want to do here is say await and then this will be an Axios response. Put that in parentheses... including around the await... then grab the data key from the response... and hydra:member from the data. That's the field on the JSON that actually holds the collection of products. Oh, and we, of course, now need to make the method async.
After this, because we only need a single featured product, I'll do a little sanity check: if featureProducts.length === 0, just return. We're not going to code too much for this case. Finish this with this.featuredProduct = featuredProducts[0].
Sweet! Wait... why is ESLint mad? Ah, it wants me to be cooler than I am and use array destructuring.