28.2. Featured Sidebar
Ok then! Put this.featuredProducts in an array and set that to featuredProducts.
That will accomplish the same thing. Now we can copy the method name, head up to created, and call it! Make sure you put it above the this.colors line: we don't want our code to await for that to finish before fetching the featured product. In this order, both AJAX requests will effectively start at the same moment.
Ok! To make sure this is working, find your shiny browser, refresh and go to the Vue Dev Tools. On Components, find ShoppingCart. Let's see... there! We do have a featuredProduct data.

The New cart-sidebar Component
The sidebar itself is going to be complex enough that I think we should put it into its own component. If you downloaded the course code, you should have a tutorial/ directory with a cart-sidebar.vue file inside.
Copy that into assets/components/shopping-cart/.
There's nothing fancy here: it has a featuredProduct required prop... and then it renders its data! Let's use this inside shopping-cart.