28.3. Featured Sidebar
You know the drill: go above the component and import CardSidebar from @/components/shopping-cart/cart-sidebar. Add that to components...
... then scroll up so we can use it. Let's see: make the aside not be self-closing... then put <cart-sidebar inside with a v-if set to featuredProduct.
I'm just coding defensively in case there is no feature product for some reason. The one prop we need to pass is :featuredProduct set to our featuredProduct data.
Awesomesauce! Move over and... there it is! The feature product is our beloved Inflatable Sofa!
The last step is to make this functional by importing and using the cart-add-controls component. Let's do that next!