29.1. Add to Cart Controls on the Sidebar
Now that we're rendering the featured product on the sidebar, our next job is to make it work! Make it possible to add the item to the cart right from this page.
Thanks to our wonderful work isolating the "add to cart" controls into its own component, this is going to be fun! Practically a victory lap!

Rendering the Cart Controls
In cart-sidebar.vue, let's get to work: import CartAddControls from @/components/product-show/cart-add-controls, add a components key, then put that inside.
Up in the template, after the h6 render <cart-add-controls. This needs a number of different props and if I hit Command+Spacebar PhpStorm will show them to me: 3 boolean props and a product object prop.
Pass :product set to featuredProduct. For the other three, hardcode them for now: I want to see this render first: :add-to-cart-loading="false", :add-to-cart-success="false" and :allow-add-to-cart="false".
Let's go check it! And... awesome! Well, not that awesome: it looks terrible! Things aren't quite... "fitting". This button really needs to be smaller.