29.2. Add to Cart Controls on the Sidebar
Adding more Flexibility to the Cart Controls Component
This is a great example of the lifecycle of a reusable component. Open cart-add-controls.vue. Until now, the "Add to Cart" text on the button did not need to be dynamic. But suddenly, we do need to be able to control it so we can make it shorter only on the sidebar.
This means our component needs to be made more configurable with a new prop. Call it addButtonText, the type is String, but this time, instead of making it required, set a default value of Add to Cart.
Back up in the template, replace the hardcoded text with {{ addButtonText }}.
Over in cart-sidebar, let's leverage this new flexibility: pass the add-button-text prop set to +. We don't need to say :add-button-text because we're not setting this to a variable or JavaScript expession: it's just a plus sign.
Over in cart-sidebar, let's leverage this new flexibility: pass the add-button-text prop set to +. We don't need to say :add-button-text because we're not setting this to a variable or JavaScript expession: it's just a plus sign.
Now when we refresh.. yes! So much better!