30.1. Deep and Smart Watchers
The fact that we can add things to our cart from the sidebar and have everything just... magically update is incredible and shows the power of.. the dark side. I mean, Vue. But... we have a bug. Gasp sounds!

We have a Bug!
Remove all of the inflatable sofa items from the cart and do a full page refresh. Now try to add a red sofa and... oh! It did not show up in the cart! I see 11 items in the cart... and I just added 1 more... so the new total should be 12... and that does show in the header! But the item is missing!

Let's check the console. Woh! Huge error!

Cannot read property price of undefined

Coming from ShoppingCartList. Yikes! The problem is... subtle. Head over to shopping-cart.vue and find our watcher function... here it is.

 130 lines  assets/pages/shopping-cart.vue
<script>
export default {
    name: 'ShoppingCart',
    watch: {
        async cart() {
            this.loadProducts();
        },
    },
};
</script>
Let's remember how this works: after the cart AJAX request finishes, Vue calls our cart watcher function and it executes the loadProducts function. That collects the product ids from the items in the cart and makes one AJAX call to fetch all of that product data. By the end of this, we have a cart object and an array that holds the data for every product in the cart.

 130 lines  assets/pages/shopping-cart.vue
<script>
export default {
    name: 'ShoppingCart',
    methods: {
        async loadProducts() {
            const productIds = this.cart.items.map((item) => item.product);
            const productsResponse = await fetchProductsById(productIds);
            this.products = productsResponse.data['hydra:member'];
        },
    },
};
</script>
The problem now is that when we add a totally new product to the cart, this, naturally, causes our Vue component to re-render. That process calls our computed property: completeCart. But this time, the new product is missing from the products data! This means the product ends up being null, which is not something we expect. Eventually, we try to read the price property from this and... well, you saw it: things fall apart.

 130 lines  assets/pages/shopping-cart.vue
<script>
export default {
    name: 'ShoppingCart',
    computed: {
        completeCart() {
            if (! this.cart || ! this.products || ! this.colors) {
                return null;
            }
            const completeItems = this.cart.items.map((cartItem) => {
                const product = this.products.find((productItem) => productItem['@id'] === cartItem.product);
                const color = this.colors.find((colorItem) => colorItem['@id'] === cartItem.color);
                return {
                    id: 'cartItem.product}_${cartItem.color ? cartItem.color : 'none'}',
                    product,
                    color,
                    quantity: cartItem.quantity,
                };
            });
            return {
                items: completeItems,
            };
        },
    },
};
</script>
But, hold on a minute... this doesn't make sense! When we add a new item to the cart, that changes the cart data. That should cause Vue to call our cart watcher function... and that should cause the load products method to make a fresh AJAX call for all of the products in the cart, including the new product.

Watcher Functions are Not "Deep"
Then, when we re-render, the completeCart computed property should work... and everything should be awesome! So... why isn't everything awesome?
The answer is that watcher functions only watch to see if an entire piece of data is changed or replaced: like when the cart goes from null to an object after the initial AJAX call. But if a piece of data is an array or an object like cart, a watcher does not watch for changes to the items in that array or the properties on that object. In other words, when a new item is added to the cart.items array, our watcher is not called. And so, the AJAX call for the fresh new products is never made!