30.2. Deep and Smart Watchers
Ok, so make the Watcher Deep!
So what's the fix? Well, you actually can make a watcher watch in "deep" mode, where it calls the function for a change on any level. Copy the loadProducts() line. A deep watcher has a different syntax. Change cart to a property set to an object with deep: true. That's the key
The function now lives under a handler() callback. Inside, paste this.loadProducts()
Let's try it! Move over to the browser, remove the sofa from the cart and refresh. Now add the green sofa back and... woohoo! The new product showed up in the cart!
But... you probably also noticed the big error in the console... which is the same error as before! That's because, after the cart changes, but before the new products AJAX call finishes, Vue re-renders and the completeCart computed prop is still missing the new product data. And so, it explodes.
A moment later, the products AJAX call finishes, Vue re-renders again, and it works fine.