30.3. Deep and Smart Watchers
Coding Defensively while we Wait for It
Vue re-renders and the completeCart computed prop is still missing the new product data. And so, it explodes.
A moment later, the products AJAX call finishes, Vue re-renders again, and it works fine.
To fix this, inside the computed prop, let's code defensively. I'm thinking that, after we go through the map function, if any of these items are missing their product, let's just filter them out. That would temporarily hide an item until its data is available.
We can do that very simply by setting the items key to completeItems.filter, with a callback that checks to see if product.item is .. truthy. So basically, this filters out any items that are missing their product.
Now, once again, refresh, remove the sofa, and refresh again. Add a red sofa with quantity 2 and... yes! That was perfect!