30.4. Deep and Smart Watchers
Watching the Cart Items Length
But... this "deep" watcher is kind of overkill. Now, even when we change the quantity of an item, it will cause a new API call for the products. Check it out: open the Network tools and filter to XHR - or AJAX requests. I'll clear things out. Each time I change the quantity, 2 AJAX calls are made! The first one saves the cart - we expect that - but the second one is because our watcher function is called... and that makes an unnecessary AJAX call for the products.
And maybe... this isn't that big of a deal. But we can do better.
So let's think about it. The only time that we really need to load the products is when a new item has been added to the cart. Could we... somehow... have a watcher watches the length of the cart.items array for changes? As strange as it sounds, we can!
Copy that loadProducts() line and change this to yet another syntax. This time, make the key a string - cart.items.length - set to a function with... any name you want, like watchCartItemsLength
That part doesn't matter. Inside, call this.loadProducts()
Yup, that works! Try it! Once again, remove the item and refresh the page so we can see the whole process. I'll add a blue sofa... oh bit first clear the requests. And... it shows up! If we change the quantity... yes! There is one AJAX call to save the updated cart, but not an extra one for the products. There is a fresh AJAX call when we remove an item, which is technically not necessary, but I am happy with this.
So... watchers are powerful, but can be tricky! That's part of the reason that I only use them when I absolutely need to. There are usually simpler, more direct solutions.
Next: I think I'm ready to buy these great products. Let's make our shopping cart page able to toggle between the cart and a checkout component. Eventually, we'll use this setup to learn all about Vue transitions.