31.1. Advancing Between Cart & Checkout
Shopping cart... check! There's just one more feature we need to add to the site: the ability to fill out a checkout form. No... we won't create a real checkout form with a credit card field and payment processing. That's not the point of this tutorial. But we will build a real form for the customer's info with real validation: both server-side and client-side.
So: should this checkout form be a new page? We could do that. But this time, instead, I'm going to add that form right to this page. This one component will hold both steps of the checkout process. Step one will show this cart and then, by clicking a button, the user will go to step two: the checkout form. This will be a great opportunity to chat about Vue transitions so that this process feels really smooth.

Adding the Checkout Button
Let's get to work. In assets/pages/shopping-cart.vue, this component renders the entire page - including cart-sidebar and shopping-cart-list.
Add a div to hold the "Checkout" button. Give it a v-if= so that we only render this once the completeCart data has finished loading and if completeCart.items.length > 0. We don't want to show a checkout button if the cart is empty.
Inside, say <button class="btn btn-primary"> and "Check Out! ".
Let's see how that looks. Nice!