31.2. Advancing Between Cart & Checkout
The "currentState" Data
Our ShoppingCart component is going to have two, sort of, "states" that the user can toggle between: the "cart" state that shows the shopping cart and the "checkout" state that shows the checkout form. We're going to need a piece of data to keep track of the current state.
Add a new currentState data that defaults to cart. I'll... make sure I spell that correctly.
Next, when the user clicks the button, let's call a new method to change that data to the other state: @click="switchState".
Copy that name and head down to the bottom of the component to add that method: switchState(). Inside, since we only have 2 states, it's pretty simple: this.currentState =, then if this.currentState === 'cart', set it to checkout, else we are in the checkout state and want to go back to cart. The button will let us go back and forth.

Vue Dev Tools Quirk when no Re-Render
We're not using currentState in the template yet... but we can at least check to see if clicking this button changes that data. Go to Vue dev tools, find ShoppingCart and find currentState. Yep, it's cart.
Click the button. Huh. It didn't change! Spoiler alert: it did change! If you click off of that component and back on... and go find that data again, it did change! This is a quirk with the Vue dev tools. If you change a piece of data, but that data doesn't cause anything to re-render. the dev tools won't instantly update. No big deal: it is working.