31.3. Advancing Between Cart & Checkout
Toggling the Title Between States
So let's render something based on the data... actually two things. When we click the button, the title should change from "Shopping Cart" to "Checkout" and the text on the button itself should go from "Check out! " to "Go Back".
Instead of hardcoding this logic and messages in the template, let's leverage some computed props. Find the computed section and call the first pageTitle. return if this.currentState === 'cart' then Shopping Cart, else Checkout
Copy this and do something similar for buttonText, returning Check Out if we're on the cart page and Back if we're on checkout. Very nice! Scroll up to use these. Change the title to :text="pageTitle" and... for the button, {{ buttonText }}
Let's check it! Back at the browser... click the button. Got it!