31.3. Advancing Between Cart & Checkout
Creating the Checkout Form Component
The last step is to toggle the content on the main part of the page. The shopping cart list is already isolated into its own component. Let's now create another component to hold the checkout form... so we don't have to add all its HTML and logic right into shopping-cart.
Create a new checkout/ directory and then an index.vue file inside. I'll paste in a very simple component that prints some text and exports a name.
Back in shopping-cart.vue, use this. Step one: import it: import CheckoutForm from @/components/checkout... but where is my auto-complete? Ah! Ryan! I accidentally put the new checkout/ directory in assets/. It should live in components/. Now Webpack is happy.
Add CheckoutForm to components and scroll up to the template.
Before we render this, look at <shopping-cart-list. Thanks to the v-if, this only renders once the completeCart data is available. Now we want to render it when the completeCart data is there and if the currentState equals cart.
Below this, render <checkout-form. Oh, but copy the v-if from above, paste, and render it when the currentState is checkout.
By the way, you could make a really good argument that the currentState logic should live in v-show instead of v-if. We'll talk more about that at the end of our discussion about Vue transitions.
Anyways, let's do this! Find your browser and click "Check Out". Beautiful! We can switch back and forth, back and forth.
But... the change is a little.. abrupt. I'd love to have a nice transition between the two pages, like the old one fades out and the new one fades in. Can we do that? Of course! CSS itself supports transitions and with a special feature from Vue, we can leverage those perfectly. Let's learn how next.