32.1. Intro to CSS Transitions
Forget about Vue for a few minutes. Instead, let's focus on how CSS transitions work by themselves. If you're already super familiar with CSS transitions, you rock. Feel free to jump to the next video. If not... let's go!

The Setup: Changing Opacity on Click
In shopping-cart.vue, add a temporary div with some text inside: testing transitions. Also use the dynamic class syntax - :class - to give this two classes: transition-testing - a class it will always have - and hidden, which it will only have when the currentState is equal to checkout.
Down in the styles, add some CSS for this: .transition-testing with opacity: 1. That's... not actually necessary... because that's the default value, but it'll help make things more clear.
Then, when .transition-testing also has the hidden class, set opacity to zero.
Basically, we're going to hide this element when we click the checkout button. Try it: there's the text and click the button. Gone! Back! Gone! Back! We're not using v-show, but the result is, effectively, the same.