32.2. Intro to CSS Transitions
CSS Transitions!
CSS transitions allow us to... well to transition between certain CSS property changes. Right now, when the hidden class is added, the opacity instantly changes from one to zero. But we can tell CSS to make that change slowly, like transitioning from one to zero over, let's say 3 seconds
To do that, on the class that the element always has, define that you want a transition to happen: transition: opacity over 3s.
Check it out: when we click, the message slooowly fades out. How cool is that! I'm old enough to remember building page layouts with tables and rounding corners with images. This... is like winning the lottery.