32.3. Intro to CSS Transitions
There are a lot more things you can do with CSS transitions, but this is the basic idea. Back in our code, remove the opacity: 1... just because it's redundant.

Switching back to v-show
So... how does this relate to life in Vue? Back up in the template, find the temporary element and change it to use a proper v-show: the way we normally hide and show things.
Copy the currentState === 'checkout' line, change this to a boring class="transition-testing", and then add v-show="", paste, but now we need the reverse: show this when currentState === 'cart'.
When we try this... we lost our transition entirely! It hides and shows instantly. We can see the reason when we inspect the element: Vue hides and shows things by adding and removing display: none. What we really need Vue to do is change the opacity from 1 to 0.
Let's talk about how Vue can do that next.