33.1. Vue Transitions
When we click the button, our little "Testing transitions" div hides and shows thanks to v-show... but without any transitions. The reason is that Vue hides and shows by adding and removing display: none. Our CSS transition will only work if the opacity changes.
So, instead of display: none, what we really need Vue to do is change the opacity from 1 to 0. Or, more generically, we need Vue to add a new class when it's hidden so that we can set opacity: 0 on that class, just like we're already doing with our hidden class. That would make this works just like before.

Hello <transition>: Adding & Removing Classes
Fortunately, Vue has a feature just for this. The key is to surround the element that we're hiding and showing with a special transition component. Check it out: add <transition before and after the div... and indent to keep ESLint happy.
But I want to be super clear about what this <transition> tag does and does not do. Go back to your browser and do a full refresh. Now, click. Woh! It doesn't even hide anymore! Oh, wait... it did hide... it just took a few seconds. When I click again, it's instantly back. That's... odd.
Inspect element on the div and watch really closely. Thanks to the <transition> component, when I click, it adds a few classes... then after a few seconds, it removes those and hides the element with display: none. When we click again, it removes the display: none and temporarily adds some other classes... which also disappear after a few seconds.
Obviously, these classes on their own do nothing: this did not magically cause the div to fade in and out. But we can use those classes in CSS to add the transitions.