33.2. Vue Transitions
<transition> Reads your CSS Transition Time
So the first takeaway is this: the <transition> component is a fancy mechanism to temporarily add a few classes whenever the item inside is hidden or shown. That's it. Though it is a bit smarter than it first appears.
Head over to the CSS in the component... and temporarily comment-out the transition: opacity 3s.
Now go back to the browser and watch the div again. When I click the button... boom! It instantly hides and does not add those temporary classes! It turns out that the transition component actually looks at the element that's being hidden or shown and detects if it has a CSS transition. If it does, it uses that to know how long to add those temporary classes.
Remove the comment and put the transition back. Now when we click, watch: the classes are added for exactly 3 seconds and then the display: none is set. Vue does this so that our CSS transition - which is not quite working yet - has the time it needs to finish before Vue finally, fully hides the element with display: none.