34.1. Transition Name & Mode
As we just learned, Vue adds, these v- classes to our element when it's being hidden or shown.
We actually have partial control over how these are named. In the template, on the transition element, add name="fade". What does that do? Very simply: it changes the prefix that Vue uses for the classes
Thanks to this, down in the styles, instead of v-, everything needs to be fade-.
We could have used name="" anything. Fade is a good name because the CSS causes the element to, ya know, fade in and out.
Move over and try it. Yep! It works like before, except that the classes are different. The advantage of this is that you could have CSS in your project for several different types of transitions, like fade or bounce and then choose whichever one you want with <transition name="".