34.2. Transition Name & Mode
Transitioning Between 2 Components
Now that we are Vue transition pros, up in the template, let's remove that temporary element. Our real goal is to transition between the shopping cart and checkout components so that it's not so... abrupt.
No problem! Wrap each component in its own transition component, right? We could do that - and I'll talk more about that approach in a few minutes. But we can also wrap a transition around multiple components or elements.
Add <transition name="fade" so that we use the same CSS, wrap this around both components and indent them.
Wrapping multiple elements or components inside a single transition only works when you're using v-if, not v-show... and it only works when there's exactly one element or component being displayed at a time... which is exactly our situation! I'll talk more about this limitation in a few minutes.
But let's try it! And... yea! It's super slow, but it works! Let's shorten that transition time to something more realistic. Down at the bottom of the component, change the 3 seconds to .2 seconds.
Try it now. Much better. Though... it's still kinda jumpy. The problem is that the new component is already being shown and fading in while the old component is still there and fading out. So it... kind of jumps a bit, which will look worse once the checkout component renders a form.