34.3. Transition Name & Mode
Transition Modes
Try it now. Much better. Though... it's still kinda jumpy. The problem is that the new component is already being shown and fading in while the old component is still there and fading out. So it... kind of jumps a bit, which will look worse once the checkout component renders a form.
Fortunately, because we wrapped both of these components inside the same transition, we can leverage a cool mode option on the transition. Say mode\out-in"
This says:

  Fade out the old component and then fade in the new one... instead of doing them at the same time.

Check it out now... awesome! The old component fades out completely before the new one starts: no jumping around.

The Problem with Transitions and v-if
But... there is one problem with the approach of wrapping multiple elements or components inside the same transition. The problem is v-if. It's not very obvious right now, but each time we click the button, the old component is completely destroyed. When we click back, it's completely re-created. That could be a slight performance issue, though that's not really the problem. The problem is that any data on the component is completely lost.
This will be much more obvious in a few minutes when we add a true checkout form. Then, if we filled out a few fields... then clicked back to the cart... then returned to the checkout form, those fields would now be blank because the component was re-created with new data and new HTML.